+++ Firmware/ScribitESP/ScribIt_wifi.cpp
@@ -307,8 +307,9 @@ void ScribIt::handleHTTPRequests()
         client.println("Content-Type: application/json");
         client.println("Access-Control-Allow-Origin: *");
         client.println();
-        client.printf("{\"state\":\"%s\",\"paused\":\"%s\",\"id\":\"%.2x%.2x%.2x%.2x%.2x%.2x\",\"version\":\"%s\"}\n",
-                      stateStr, pauseStr, m_ID[0], m_ID[1], m_ID[2], m_ID[3], m_ID[4], m_ID[5], FIRMWARE_VERSION);
+        client.printf("{\"state\":\"%s\",\"paused\":\"%s\",\"id\":\"%.2x%.2x%.2x%.2x%.2x%.2x\",\"version\":\"%s\",\"samd\":\"%s\"}\n",
+                      stateStr, pauseStr, m_ID[0], m_ID[1], m_ID[2], m_ID[3], m_ID[4], m_ID[5], FIRMWARE_VERSION,
+                      m_samdSynced ? "ok" : "not_detected");
     }
     else if (path == "/upload" && method == "POST")
     {
@@ -454,10 +455,11 @@ void ScribIt::handleHTTPRequests()
         client.println("Access-Control-Allow-Origin: *");
         client.println();
         client.printf(
-            "{\"state\":\"%s\",\"paused\":\"%s\",\"id\":\"%.2x%.2x%.2x%.2x%.2x%.2x\",\"version\":\"%s\"}\\n",
+            "{\"state\":\"%s\",\"paused\":\"%s\",\"id\":\"%.2x%.2x%.2x%.2x%.2x%.2x\",\"version\":\"%s\",\"samd\":\"%s\"}\\n",
             stateStr, pauseStr,
             m_ID[0], m_ID[1], m_ID[2], m_ID[3], m_ID[4], m_ID[5],
-            FIRMWARE_VERSION);
+            FIRMWARE_VERSION,
+            m_samdSynced ? "ok" : "not_detected");
     }
     else if (path == "/gcode" && method == "POST")
     {